style: aligning patient page style with design
