+++ cypress/e2e/components/paginator.cy.ts
@@ -1,0 +1,6 @@
+describe('Paginator test', () => {
+  it('Paginator is visible', () => {
+    cy.visit('/pacientes')
+    cy.get('[data-testid="paginator"]').should('be.visible')
+  })
+})
