+++ cypress/e2e/page/patients/patients.cy.ts
@@ -18,36 +18,6 @@ describe('Patients page test', () => {
     cy.get('[data-testid="searchSection"]').find('p').should('be.visible')
   })
 
-  it('Patients data section loaded', () => {
-    cy.get('[data-testid="patients-data"]').should('be.visible')
-  })
-
-  it('Patients view filter is loaded', () => {
-    cy.get('[data-testid="patients-data"]')
-      .find('main')
-      .find('div')
-      .should('be.visible')
-  })
-
-  it('Patient view is toggling properly', () => {
-    cy.get('[data-testid="patients-data"]')
-      .find('main>div>button:first')
-      .should('have.class', '_isActive_1d6sd_72')
-    cy.get('[data-testid="patients-data"]')
-      .find('main>div>button:nth-child(2)')
-      .should('not.have.class', '_isActive_1d6sd_72')
-
-    cy.get('[data-testid="patients-data"]')
-      .find('main>div>button:nth-child(2)')
-      .click()
-
-    cy.get('[data-testid="patients-data"]')
-      .find('main>div>button:first')
-      .should('not.have.class', '_isActive_1d6sd_72')
-    cy.get('[data-testid="patients-data"]')
-      .find('main>div>button:nth-child(2)')
-      .should('have.class', '_isActive_1d6sd_72')
-  })
   /**
    * TODO: Create a test for labels when this are created dynamically from the filters menu selected options
    */
