+++ src/components/ui/label/Label.module.css
@@ -1,15 +1,16 @@
 .label {
   display: flex;
-  align-items: end;
+  align-items: center;
   justify-content: center;
+  height: 26px;
   gap: 8px;
   background-color: var(--container-color);
   border: 1px solid var(--border-color);
   border-radius: var(--border-radius-md);
-  padding: 8px 12px;
+  padding: 4px 8px;
 
   p {
-    font-size: 14px;
+    font-size: 12px;
     font-weight: 500;
     color: var(--primary-text);
   }
