+++ src/components/ui/label/Label.tsx
@@ -3,14 +3,15 @@ import { getIcon } from '@/utils/getIcons'
 
 interface Props {
   text: string
+  customStyle?: object
   color?: string
   icon?: string
 }
 
-export function Label({ text, color = '#fff', icon = '' }: Props) {
+export function Label({ text, customStyle, color = '#fff', icon = '' }: Props) {
   return (
     <>
-      <div className={styles.label}>
+      <div className={styles.label} style={customStyle}>
         <p> {text} </p>
         {getIcon(icon, { color })}
       </div>
