+++ src/layouts/dashboard/DashBoardView.module.css
@@ -11,4 +11,5 @@
 .main {
   grid-area: main;
   padding: 60px;
+  padding-bottom: 33px;
 }
