+++ src/pages/patients/Patients.module.css
@@ -43,175 +43,3 @@
   display: flex;
   gap: 16px;
 }
-
-.patients-data {
-  display: grid;
-  grid-template-columns: 1000px 526px;
-  grid-template-rows: 559px 32px;
-  justify-items: center;
-  column-gap: 24px;
-  padding: 26px 0;
-  width: 100%;
-
-  & footer {
-    display: flex;
-    align-items: center;
-    width: max-content;
-    margin-top: 16px;
-
-    & > div {
-      &:not(:nth-child(2)) {
-        background-color: var(--container-color);
-        border-radius: 50%;
-        padding: 8px;
-
-        & svg {
-          width: 16px;
-        }
-      }
-
-      &:nth-child(2) {
-        display: flex;
-        justify-content: space-aroun;
-        gap: 12px;
-        margin: 0 16px;
-        font-size: 16px;
-        font-weight: 500;
-
-        & p:first-child {
-          color: var(--accent-color);
-          font-size: 24px;
-          font-weight: 700;
-        }
-      }
-    }
-  }
-
-  & .patients-selection {
-    display: flex;
-    flex-direction: column;
-    background-color: inherit;
-
-    & > div:first-child {
-      width: max-content;
-      padding: 5px;
-      background: var(--container-color);
-      border-radius: var(--border-radius-md);
-
-      & > button {
-        padding: 6px 35px;
-        border-radius: var(--border-radius-sm);
-        font-size: 14px;
-        font-weight: 500;
-
-        &:hover,
-        &.isActive {
-          background-color: var(--border-color);
-        }
-      }
-    }
-
-    & > div:nth-child(2) div {
-      display: flex;
-      justify-content: space-between;
-      margin: 10px 0;
-
-      & > div {
-        display: flex;
-        align-items: center;
-        gap: 5px;
-        font-size: 16px;
-        font-weight: 500;
-      }
-
-      & select {
-        color: var(--input-text);
-        background-color: var(--input-bg-color);
-      }
-
-      & svg {
-        width: 24px;
-        height: 40px;
-        cursor: pointer;
-      }
-    }
-
-    & .patients-table {
-      height: 460px;
-
-      & tbody {
-        display: block;
-        overflow-y: scroll;
-        height: 412px;
-        -ms-overflow-style: none; /* IE and Edge */
-        scrollbar-width: none; /* Firefox */
-
-        &::-webkit-scrollbar {
-          display: none; /* Chrome */
-        }
-      }
-
-      & tr {
-        display: grid;
-        justify-items: center;
-        grid-template-columns: 140px repeat(3, 218px) 140px;
-        background-color: var(--container-color);
-        border: 1px solid var(--border-color);
-        border-radius: var(--border-radius-md);
-        font-size: 16px;
-        font-weight: 500;
-        padding: 9px 32px;
-
-        tbody & {
-          margin-top: 8px;
-          cursor: pointer;
-
-          & td:last-child {
-            justify-self: end;
-
-            & > svg {
-              visibility: hidden;
-            }
-          }
-
-          &:hover {
-            border: 1px solid var(--accent-color);
-
-            & td:last-child > svg {
-              visibility: visible;
-            }
-          }
-        }
-      }
-
-      & td {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        gap: 5px;
-
-        tbody &:not(:first-child) {
-          color: var(--secondary-text);
-        }
-
-        thead & svg {
-          width: 15px;
-          cursor: pointer;
-        }
-      }
-    }
-  }
-
-  & > aside {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 12px;
-    background-color: var(--container-color);
-    border: 1px solid var(--border-color);
-    border-radius: var(--border-radius-md);
-    width: 100%;
-    height: 100%;
-  }
-}
