+++ src/pages/patients/components/PatientsTable/objectives.tsx
@@ -1,0 +1,10 @@
+const objectives: { [index: string]: string } = {
+  'Perder peso': '#007A4D',
+  Musculatura: '#6D6D6D',
+  'Controlar diabetes': '#007772',
+  Crecimiento: '#0265DC',
+  'Aumentar de peso': '#5258E4',
+  'Ensanchar caderas': '#893DE7',
+}
+
+export default objectives
