+++ src/pages/patients/components/PatientsTable/patients-table.json
@@ -3,60 +3,60 @@
     "name": "Andres",
     "email": "andres@gmail.com",
     "lastAppointment": "01/01/2025 a las 12:34pm",
-    "objective": "Objetivo 1"
+    "objective": "Perder peso"
   },
   {
     "name": "Andres",
     "email": "andres@gmail.com",
     "lastAppointment": "01/01/2025 a las 12:34pm",
-    "objective": "Objetivo 2"
+    "objective": "Musculatura"
   },
   {
     "name": "Andres",
     "email": "andres@gmail.com",
     "lastAppointment": "01/01/2025 a las 12:34pm",
-    "objective": "Objetivo 3"
+    "objective": "Controlar diabetes"
   },
   {
     "name": "Andres",
     "email": "andres@gmail.com",
     "lastAppointment": "01/01/2025 a las 12:34pm",
-    "objective": "Objetivo 4"
+    "objective": "Crecimiento"
   },
   {
     "name": "Andres",
     "email": "andres@gmail.com",
     "lastAppointment": "01/01/2025 a las 12:34pm",
-    "objective": "Objetivo 5"
+    "objective": "Aumentar de peso"
   },
   {
     "name": "Andres",
     "email": "andres@gmail.com",
     "lastAppointment": "01/01/2025 a las 12:34pm",
-    "objective": "Objetivo 6"
+    "objective": "Ensanchar caderas"
   },
   {
     "name": "Andres",
     "email": "andres@gmail.com",
     "lastAppointment": "01/01/2025 a las 12:34pm",
-    "objective": "Objetivo 7"
+    "objective": "Crecimiento"
   },
   {
     "name": "Andres",
     "email": "andres@gmail.com",
     "lastAppointment": "01/01/2025 a las 12:34pm",
-    "objective": "Objetivo 8"
+    "objective": "Ensanchar caderas"
   },
   {
     "name": "Andres",
     "email": "andres@gmail.com",
     "lastAppointment": "01/01/2025 a las 12:34pm",
-    "objective": "Objetivo 9"
+    "objective": "Controlar diabetes"
   },
   {
     "name": "Andres",
     "email": "andres@gmail.com",
     "lastAppointment": "01/01/2025 a las 12:34pm",
-    "objective": "Objetivo 10"
+    "objective": "Musculatura"
   }
 ]
