+++ src/pages/patients/components/RegisterModal/RegisterModal.tsx
