+++ src/pages/patients/components/RegisterModal/register-inputs.json
